ml-model/requirements.txt
flask==2.2.2
Werkzeug==2.2.2
requests==2.28.1
PyJWT==2.6.0
cryptography==38.0.4